+++ packages/f-alert/CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+Latest (to be added to next release)
+------------------------------
+*October 23, 2020*
+
+### Added
+- Stylelint added to lint styling on build.
+- data-test IDs to Alert Component 
+- Basic tests for Alert 
+- Alert Component-Object 
+
+
 v0.1.1
 ------------------------------
 *October 20, 2020*